Remove `webpki::TrustAnchor` from the public API.
The `webpki::TrustAnchor` type will change significantly in the near
future. Change Rustls's public APIs so that we can upgrade to the new
version of webpki without it being a breaking change. That is, allow
webpki and Rustls to evolve at different paces.

We expect that a future version of `webpki-roots` will also remove the
`webpki::TrustAnchor` from its public API, replacing it with its own
similar type.

The new `from_subject_spki_name_constraints` constructor will allow such
refactorings to be done without having to create a separate crate to
contain the types.

This makes the use of webpki-roots a bit messier than now, but most
applications probably shouldn't be using webpki-roots anyway.