docs: add rustdoc for DigitallySignedStruct.
The front-page index for the Rustls lib includes a list of re-exported
structs comprising the public API. Prior to this commit the
`handskae::DigitallySignedStruct` type was included in this list, but
lacked a rustdoc comment.

This commit adds a simple rustdoc comment for this type. Now all of the
structs mentioned in the front-page index have some documentation.