feat: impl safer export_keyring_material interface.
Prior to this commit the `export_keyring_material` function used
a mutable out buffer for writing exported key material, and returned an
empty Ok result when there was no error doing so.

This commit updates the function such that the ownership of the output
buffer passes through the export function and is returned as the Ok
result when there is no error.

Doing this makes for a safer interface for end users: the output buffer
will be dropped if `export_keyring_material` errors. Callers can only
access the buffer again using the OK result.

All credit due to davidv1992 for the implementation idea and initial
code that was extended in this commit.