CI: add fmt and clippy coverage for fuzz workspace.
The `fuzz` subdirectory is set up as a separate workspace from the main
workspace that contains `rustls` and `examples`. Because of this running
`cargo fmt --all` and other similar tooling doesn't include the `fuzz`
directory. This can lead to formatting/clippy drift over time.

In this commit the `build.yml` config is extended to also run `clippy`
and `fmt` on the `fuzz` subdirectory using `--manifest-path`.